back:feat: add activity deleting verification

+++ doggienote-api/src/dict-activity/dict-activity.service.ts
@@ -2,14 +2,16 @@ import { Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { DictActivity } from './dic-activity.entity';
+import { ActivityService } from 'src/activity/activity.service';
 
 @Injectable()
 export class DictActivityService {
   private readonly logger = new Logger(DictActivityService.name);
-  
+
   constructor(
     @InjectRepository(DictActivity)
     private dictActivityRepository: Repository<DictActivity>,
+    private activityServis: ActivityService,
   ) {}
 
   async findAll(): Promise<DictActivity[]> {
@@ -21,7 +23,7 @@ export class DictActivityService {
   }
 
   async createDictActivity(
-    dictActivityData: Partial<DictActivity>
+    dictActivityData: Partial<DictActivity>,
   ): Promise<DictActivity> {
     const newDictActivity =
       this.dictActivityRepository.create(dictActivityData);
@@ -41,16 +43,29 @@ export class DictActivityService {
     return this.dictActivityRepository.save(updatedDictActivity);
   }
 
-  async deleteActivity(id: string) {
-    const dictActivityToDelete =
-      await this.dictActivityRepository.findOneOrFail({ where: { id } });
-    if (dictActivityToDelete.removable === true) {
-      // const dictActivityNameDeleted = dictActivityToDelete.dict_activity;
-      await this.dictActivityRepository.delete(id);
-      //TODO: handle returning deleted dict_activity name
-      this.logger.log(`DictActivitY: ${dictActivityToDelete.id} was deleted`);
-    } else {
-      return new Error('This activity cannot be deleted.');
+  async deleteActivity(id: string): Promise<Error> {
+    try {
+      const dictActivityToDelete =
+        await this.dictActivityRepository.findOneOrFail({ where: { id } });
+
+      if (dictActivityToDelete.removable === true) {
+
+        const activity = await this.activityServis.findByIdDictActivity(id);
+
+        if (activity.length === 0) {
+          await this.dictActivityRepository.delete(id);
+          //TODO: handle returning deleted dict_activity name
+          this.logger.log(`DictActivitY: ${dictActivityToDelete.id} was deleted`);
+        } else {
+          throw new Error(
+            'This dictActivity is used in Activity database. It cannot be removed.',
+          );
+        }
+      } else {
+        return new Error('This activity cannot be deleted.');
+      }
+    } catch (error) {
+      this.logger.error(error);
     }
   }
 }
